Move generator-output checks into the verification module
