+++ html/kotlinx-browser-common-subset/generator/verification/src/jvmTest/kotlin/GeneratedClosureTest.kt
