+++ html/kotlinx-browser-common-subset/generator/verification/src/jvmTest/kotlin/GeneratedEnumLikeValueLedgerTest.kt
